move conftest to top level
