+++ tests/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+from tests.extension.mockextensions.app import MockExtensionApp
+
 
 pytest_plugins = ["jupyter_server.pytest_plugin"]
 
@@ -29,3 +31,51 @@ def pytest_runtest_setup(item):
             pytest.skip(
                 "Skipping this test because it's marked 'integration_test'. Run integration tests using the `--integration_tests` flag."
             )
+
+
+mock_html = """
+<!DOCTYPE HTML>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>{% block title %}Jupyter Server 1{% endblock %}</title>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    {% block meta %}
+    {% endblock %}
+</head>
+<body>
+  <div id="site">
+    {% block site %}
+    {% endblock site %}
+  </div>
+  {% block after_site %}
+  {% endblock after_site %}
+</body>
+</html>
+"""
+
+
+@pytest.fixture
+def mock_template(jp_template_dir):
+    index = jp_template_dir.joinpath("index.html")
+    index.write_text(mock_html)
+
+
+@pytest.fixture
+def extension_manager(jp_serverapp):
+    return jp_serverapp.extension_manager
+
+
+@pytest.fixture
+def config_file(jp_config_dir):
+    """"""
+    f = jp_config_dir.joinpath("jupyter_mockextension_config.py")
+    f.write_text("c.MockExtensionApp.mock_trait ='config from file'")
+    return f
+
+
+@pytest.fixture(autouse=True)
+def jp_mockextension_cleanup():
+    yield
+    MockExtensionApp.clear_instance()
